requirements.txt
flask
gunicorn
openai>=1.0.0
python-dotenv
pandas
nltk
faiss-cpu
numpy
flask_cors
transformers
torch
scikit-learngit